# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Yüz Tanıma ile Yoklama Sistemi API",
    "description": "Yüz tanıma teknolojisi kullanarak yoklama almayı ve duygu analizi yapmayı amaçlayan bir sistemin API'si",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
...
  "paths": {
...
    "/api/attendance/course/{course_id}": {
      "post": {
        "tags": [
          "✅ Yoklama İşlemleri"
        ],
        "summary": "Yoklama Al",
        "description": "Sınıf fotoğrafı ile yoklama alır ve yüz tanıma yapar",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Ders ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "photo",
            "in": "formData",
            "description": "Sınıf fotoğrafı",
            "required": true,
            "type": "file"
          },
          {
            "name": "lesson_number",
            "in": "formData",
            "description": "Ders saati numarası",
            "required": false,
            "type": "integer",
            "default": 1
          }
        ],
        "responses": {
          "201": {
            "description": "Yoklama başarıyla alındı",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Yoklama başarıyla alındı."
                },
                "attendance": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz istek",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu ders için bugün zaten yoklama alınmış."
                }
              }
            }
          },
          "403": {
            "description": "Yetki hatası",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu işlem için yetkiniz yok."
                }
              }
            }
          },
          "404": {
            "description": "Ders bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Ders bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Yoklama alınamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Yoklama alınamadı."
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "✅ Yoklama İşlemleri"
        ],
        "summary": "Dersin yoklamalarını getir",
        "description": "Belirli bir dersin yoklama kayıtlarını getirir",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Ders ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Başlangıç tarihi (YYYY-MM-DD)",
            "required": false,
            "type": "string"
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Bitiş tarihi (YYYY-MM-DD)",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Sayfa numarası",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Sayfa başına kayıt sayısı",
            "required": false,
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "page": {
                  "type": "integer",
                  "example": 1
                },
                "per_page": {
                  "type": "integer",
                  "example": 10
                },
                "total": {
                  "type": "integer",
                  "example": 5
                },
                "pages": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          },
          "403": {
            "description": "Yetki hatası",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu dersin yoklamalarını görüntüleme yetkiniz yok."
                }
              }
            }
          },
          "404": {
            "description": "Ders bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Ders bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Yoklamalar getirilemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Yoklamalar getirilemedi."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/attendance/course/{course_id}": {
      "get": {
        "tags": [
          "📊 Yoklama Raporları"
        ],
        "summary": "Ders bazlı devamsızlık raporu",
        "description": "Bir dersteki genel devamsızlık oranını ve öğrenci bazında istatistikleri gösterir",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Ders ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Başlangıç tarihi (YYYY-MM-DD)",
            "required": false,
            "type": "string"
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Bitiş tarihi (YYYY-MM-DD)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "course": {
                  "type": "object"
                },
                "date_range": {
                  "type": "object",
                  "properties": {
                    "start_date": {
                      "type": "string",
                      "example": "2024-04-01"
                    },
                    "end_date": {
                      "type": "string",
                      "example": "2024-04-30"
                    }
                  }
                },
                "statistics": {
                  "type": "object",
                  "properties": {
                    "total_attendances": {
                      "type": "integer",
                      "example": 10
                    },
                    "total_students": {
                      "type": "integer",
                      "example": 20
                    },
                    "total_records": {
                      "type": "integer",
                      "example": 200
                    },
                    "present_count": {
                      "type": "integer",
                      "example": 150
                    },
                    "absent_count": {
                      "type": "integer",
                      "example": 30
                    },
                    "late_count": {
                      "type": "integer",
                      "example": 15
                    },
                    "excused_count": {
                      "type": "integer",
                      "example": 5
                    },
                    "attendance_rate": {
                      "type": "number",
                      "example": 82.5
                    }
                  }
                },
                "student_statistics": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "attendances": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Yetki hatası",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu dersin raporlarını görüntüleme yetkiniz yok."
                }
              }
            }
          },
          "404": {
            "description": "Ders bulunamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Ders bulunamadı."
                }
              }
            }
          },
          "500": {
            "description": "Rapor oluşturulurken bir hata oluştu",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Rapor oluşturulurken bir hata oluştu."
                }
              }
            }
          }
        }
      }
    },
    "/api/courses": {
      "get": {
        "tags": [
          "📚 Dersler"
        ],
        "summary": "Tüm dersleri listele",
        "description": "Sistemdeki tüm dersleri listeler",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Sayfa numarası",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Sayfa başına kayıt sayısı",
            "required": false,
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "page": {
                  "type": "integer",
                  "example": 1
                },
                "per_page": {
                  "type": "integer",
                  "example": 10
                },
                "total": {
                  "type": "integer",
                  "example": 5
                },
                "pages": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          },
          "403": {
            "description": "Bu işlem için yetkiniz yok",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu işlem için yetkiniz yok."
                }
              }
            }
          },
          "500": {
            "description": "Dersler listelenemedi",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Dersler listelenemedi."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "📚 Dersler"
        ],
        "summary": "Yeni ders oluştur",
        "description": "Sisteme yeni bir ders ekler",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ders bilgileri",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "code",
                "name",
                "semester",
                "teacher_id",
                "lesson_times"
              ],
              "properties": {
                "code": {
                  "type": "string",
                  "example": "CSE101"
                },
                "name": {
                  "type": "string",
                  "example": "Bilgisayar Mühendisliğine Giriş"
                },
                "semester": {
                  "type": "string",
                  "example": "2023-BAHAR"
                },
                "teacher_id": {
                  "type": "integer",
                  "example": 1
                },
                "lesson_times": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "day": {
                        "type": "string",
                        "example": "MONDAY"
                      },
                      "start_time": {
                        "type": "string",
                        "example": "09:00"
                      },
                      "end_time": {
                        "type": "string",
                        "example": "09:50"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ders başarıyla oluşturuldu",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Geçersiz istek",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "En az bir ders saati gerekli."
                }
              }
            }
          },
          "403": {
            "description": "Bu işlem için öğretmen yetkisi gerekli",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu işlem için öğretmen yetkisi gerekli."
                }
              }
            }
          },
          "500": {
            "description": "Ders oluşturulamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Ders oluşturulamadı."
                }
              }
            }
          }
        }
      }
    },
...
    "/api/courses/{course_id}/students/{student_id}": {
      "delete": {
        "tags": [
          "📚 Dersler"
        ],
        "summary": "Dersten öğrenci çıkar",
        "description": "Belirli bir dersten öğrenciyi çıkarır",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Ders ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "student_id",
            "in": "path",
            "description": "Öğrenci ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Öğrenci başarıyla çıkarıldı"
          },
          "403": {
            "description": "Bu işlem için yetkiniz yok",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bu işlem için yetkiniz yok."
                }
              }
            }
          },
          "404": {
            "description": "Ders, öğrenci bulunamadı veya öğrenci bu derste kayıtlı değil",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğrenci bu derse kayıtlı değil."
                }
              }
            }
          },
          "500": {
            "description": "Öğrenci çıkarılamadı",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Öğrenci çıkarılamadı."
                }
              }
            }
          }
        }
      }
    },
...
  }
} 